fix: JS to TS conversion of scripts

+++ src/app/api/income/[id].ts
@@ -1,15 +1,17 @@
+import { NextApiRequest, NextApiResponse } from "next";
 import prisma from "../../../prisma/client";
 
-export default async function handler(req, res) {
+export default async function handler(req:NextApiRequest, res: NextApiResponse) {
 
   try {
     if (req.method === "GET") {
       // Fetch income by ID
-      const state = req.param['state'];
-      const local = req.param['local'];
-      const ministry = req.param['ministry'];
-      const id = req.param['id'];
+      const state = req.query['state'] as string;
+      const localParam = req.query['local'] as string;
+      const ministry = req.query['ministry'] as string;
+      const id = req.query['id'] as string;
 
+      
       if(id){
         const income = await prisma.income.findUnique({
           where: { iid: id },
@@ -40,24 +42,24 @@ export default async function handler(req, res) {
         return res.status(200).json(income);
       }
 
-      if(local){
+      if(localParam){
         const local = await prisma.local.findMany({
-          where: { local: local },
+          where: { localId: localParam },
         });
         if (!local) {
           return res.status(404).json({ error: "Local not found" });
         }
-        return res.status(200).json(local);
+        return res.status(200).json(localParam);
       }
 
       if(state){
-        const state = await prisma.state.findMany({
+        const stateRes = await prisma.state.findMany({
           where: { state: state },
         });
-        if (!state) {
+        if (!stateRes) {
           return res.status(404).json({ error: "State not found" });
         }
-        return res.status(200).json(state);
+        return res.status(200).json(stateRes);
       }
 
 
@@ -65,19 +67,23 @@ export default async function handler(req, res) {
       // Update income by ID
       const { name, desc, date, t_amount, expenditure, central, state, local } =
         req.body;
+        const id = req.query["id"] as string;
 
+        if (!id) {
+          return res.status(400).json({ error: "Ministry ID is required" });
+        }
       const updatedIncome = await prisma.income.update({
         where: { iid: id },
         data: {
           name,
           desc,
-          date: date ? new Date(date) : undefined,
+          date: date,
           t_amount,
           central,
           state: state ? { connect: { sid: state } } : undefined,
           local: local ? { connect: { lid: local } } : undefined,
           expenditure: expenditure
-            ? { connect: expenditure.map((eid) => ({ eid })) }
+            ? { connect: expenditure.map((eid: string) => ({ eid })) }
             : undefined,
         },
       });
@@ -85,6 +91,11 @@ export default async function handler(req, res) {
       return res.status(200).json(updatedIncome);
     } else if (req.method === "DELETE") {
       // Delete income by ID
+      const id = req.query["id"] as string;
+
+      if (!id) {
+        return res.status(400).json({ error: "Ministry ID is required" });
+      }
       await prisma.income.delete({
         where: { iid: id },
       });
